% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSMNmoments.R
\name{TSMNmoments}
\alias{TSMNmoments}
\title{Moments of Truncated Scale Mixtures of Normal Distributions}
\usage{
TSMNmoments(mu, sigma2, nu = NULL, delta = NULL, lower = -Inf,
  upper = Inf, dist = "Normal")
}
\arguments{
\item{mu}{Location parameter.}

\item{sigma2}{Scale parameter.}

\item{nu}{Parameter of the scale variable of the SMN family. Must be NULL in
case of Normal distribution. Must be a bidimensional vector and each
component of the bidimensional vector "nu" must lie on (0,1) in case of
contaminated normal distribution (CNormal).}

\item{delta}{Second parameter of Pearson VII. Must not be provided in case of
Normal, Student-t or Slash distribution.}

\item{lower}{Lower bounds.}

\item{upper}{Upper bounds.}

\item{dist}{Distribution to be used: "Normal" for Normal model, "T" for
Student-t model, "PearsonVII" for Pearson VII model, "Slash" for slash
model and "CNormal" for contaminated Normal model.}
}
\value{
\describe{ \item{EY1}{The first moment} \item{EY2}{The second moment}
  \item{EY3}{The third moment} \item{EY4}{The fourth moment} }
}
\description{
\code{TSMNmoments} return the first four moments of TSMN distributions, as
example Normal, Student-t, Pearson VII, Slash or Contaminated Normal.
}
\examples{
 ## A test sample to compare theoretical and empirical moments, considering the following parameters:

 mu = 2
 sigma2 = 4
 nu = 5
 lower = -3
 upper = 10
 dist = "T"
 n = 10000

 ## Theoretical moments with TSMNmoments
 theor<-TSMNmoments(mu=mu, sigma2=sigma2, nu=nu, lower=lower, upper=upper, dist=dist)

 ## Generate the sample with TSMNgenerator to compute the empirical moments
 empir<-TSMNgenerator(n=n, mu=mu, sigma2=sigma2, nu=nu, lower=lower, upper=upper, dist=dist)

 ## Compare the results
 data.frame("1st" = c("Theoretic" = theor$EY1, "Empirical" = mean(empir)),
            "2nd" = c("Theoretic" = theor$EY2, "Empirical" = mean(empir^2)),
            "3rd" = c("Theoretic" = theor$EY3, "Empirical" = mean(empir^3)),
            "4th" = c("Theoretic" = theor$EY4, "Empirical" = mean(empir^4)))
}
\seealso{
\code{\link{TSMNgenerator}}
}
